fix: Custom hook works on reload page
