+++ services/web-app/src/app/lib/hooks/useAxios.tsx
@@ -1,35 +1,12 @@
-import { useState, useEffect } from "react";
-import axios from "axios";
+import axios, { AxiosInstance } from "axios";
 import { BASE_URL } from "@/app/lib/constants";
-import { AxiosRequest } from "@/app/lib/types";
 
-const useAxios = (request: AxiosRequest) => {
-  const [data, setData] = useState(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState(null);
+const axiosInstance = axios.create({
+  baseURL: `${BASE_URL}`,
+});
 
-  const fetchData = async () => {
-    setLoading(true);
-    setError(null);
-
-    try {
-      const response = await axios[request.method](BASE_URL + request.path, {
-        params: request.params ?? undefined,
-        data: request.body ?? undefined,
-      });
-      setData(response.data);
-    } catch (err: any) {
-      setError(err);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchData();
-  }, []);
-
-  return { data, loading, error };
+const useAxios = (): { axiosInstance: AxiosInstance } => {
+  return { axiosInstance };
 };
 
 export default useAxios;
